Throw more http exceptions on bad requests

+++ src/idea/idea.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, HttpException, HttpStatus } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 
@@ -23,16 +23,23 @@ export class IdeaService {
   }
 
   async read(id:string) {
-    return await this.ideaRepository.findOne({ where: { id } });
+    const idea = await this.ideaRepository.findOne({ where: { id } });
+    if (!idea) throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
+    return idea;
   }
 
   async update(id:string, data:Partial<IdeaDTO>) {
+    const idea = await this.ideaRepository.findOne({ where: { id } });
+    if (!idea) throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
     await this.ideaRepository.update({ id }, data);
-    return await this.ideaRepository.findOne({ id });
+    return idea;
   }
 
   async destroy(id:string) {
+    const idea = await this.ideaRepository.findOne({ where: { id } });
+    if (!idea) throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
     await this.ideaRepository.delete({ id });
-    return { deleted: true };
+    return idea;
   }
+
 }
